Выделит логику карточки фильма в отдельный презентер
